requirements.txt
apify-client>=1.6.0
python-dotenv>=1.0.0
playwright>=1.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
pyinstaller>=6.0.0
